Move creation of builtin writers to the participant factory
