+++ dds/src/implementation/rtps/discovery_types.rs
@@ -3,6 +3,32 @@
 
 use crate::serialized_payload::cdr::{deserialize::CdrDeserialize, serialize::CdrSerialize};
 
+use super::types::{EntityId, BUILT_IN_READER_WITH_KEY, BUILT_IN_WRITER_WITH_KEY};
+
+pub const ENTITYID_SPDP_BUILTIN_PARTICIPANT_WRITER: EntityId =
+    EntityId::new([0x00, 0x01, 0x00], BUILT_IN_WRITER_WITH_KEY);
+
+pub const ENTITYID_SPDP_BUILTIN_PARTICIPANT_READER: EntityId =
+    EntityId::new([0x00, 0x01, 0x00], BUILT_IN_READER_WITH_KEY);
+
+pub const ENTITYID_SEDP_BUILTIN_TOPICS_ANNOUNCER: EntityId =
+    EntityId::new([0, 0, 0x02], BUILT_IN_WRITER_WITH_KEY);
+
+pub const ENTITYID_SEDP_BUILTIN_TOPICS_DETECTOR: EntityId =
+    EntityId::new([0, 0, 0x02], BUILT_IN_READER_WITH_KEY);
+
+pub const ENTITYID_SEDP_BUILTIN_PUBLICATIONS_ANNOUNCER: EntityId =
+    EntityId::new([0, 0, 0x03], BUILT_IN_WRITER_WITH_KEY);
+
+pub const ENTITYID_SEDP_BUILTIN_PUBLICATIONS_DETECTOR: EntityId =
+    EntityId::new([0, 0, 0x03], BUILT_IN_READER_WITH_KEY);
+
+pub const ENTITYID_SEDP_BUILTIN_SUBSCRIPTIONS_ANNOUNCER: EntityId =
+    EntityId::new([0, 0, 0x04], BUILT_IN_WRITER_WITH_KEY);
+
+pub const ENTITYID_SEDP_BUILTIN_SUBSCRIPTIONS_DETECTOR: EntityId =
+    EntityId::new([0, 0, 0x04], BUILT_IN_READER_WITH_KEY);
+
 #[derive(PartialEq, Eq, Debug, Clone, Copy, CdrSerialize, CdrDeserialize)]
 pub struct BuiltinEndpointSet(pub u32);
